Pick the random example on the server-side and initially select the "main" file
Before this, the server and client could pick different random examples,
causing a mis-match render on the client-side. This was because the
random pick was happening in the `constructor()` instead of
`getInitialProps()`. Now fixed.

Also there's a concept of "main" file that is bundled into the
`now-examples.json` data file at build-time, so that the correct main
entrypoint file is initially selected in the code editor.